+++ familyhelper-finance-sdk/src/main/java/com/dwarfeng/familyhelper/finance/sdk/bean/dto/WebInputAccountBookUpdateInfo.java
@@ -34,6 +34,7 @@ public class WebInputAccountBookUpdateInfo implements Dto {
 
     @JSONField(name = "account_book_key")
     @Valid
+    @NotNull
     private WebInputLongIdKey accountBookKey;
 
     @JSONField(name = "name")