+++ familyhelper-finance-sdk/src/main/java/com/dwarfeng/familyhelper/finance/sdk/bean/dto/WebInputFundChangeUpdateInfo.java
@@ -2,8 +2,10 @@ package com.dwarfeng.familyhelper.finance.sdk.bean.dto;
 
 import com.alibaba.fastjson.annotation.JSONField;
 import com.dwarfeng.familyhelper.finance.stack.bean.dto.FundChangeUpdateInfo;
+import com.dwarfeng.subgrade.sdk.bean.key.WebInputLongIdKey;
 import com.dwarfeng.subgrade.stack.bean.dto.Dto;
 
+import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 import java.math.BigDecimal;
 import java.util.Objects;
@@ -16,19 +18,25 @@ import java.util.Objects;
  */
 public class WebInputFundChangeUpdateInfo implements Dto {
 
-    private static final long serialVersionUID = 7873346983022175175L;
+    private static final long serialVersionUID = 6862927934191227195L;
 
     public static FundChangeUpdateInfo toStackBean(WebInputFundChangeUpdateInfo webInputFundChangeUpdateInfo) {
         if (Objects.isNull(webInputFundChangeUpdateInfo)) {
             return null;
         } else {
             return new FundChangeUpdateInfo(
+                    WebInputLongIdKey.toStackBean(webInputFundChangeUpdateInfo.getFundChangeKey()),
                     webInputFundChangeUpdateInfo.getDelta(),
                     webInputFundChangeUpdateInfo.getChangeType(),
                     webInputFundChangeUpdateInfo.getRemark()
             );
         }
     }
+
+    @JSONField(name = "fund_change_key")
+    @Valid
+    @NotNull
+    private WebInputLongIdKey fundChangeKey;
 
     @JSONField(name = "delta")
     @NotNull
@@ -41,6 +49,14 @@ public class WebInputFundChangeUpdateInfo implements Dto {
     private String remark;
 
     public WebInputFundChangeUpdateInfo() {
+    }
+
+    public WebInputLongIdKey getFundChangeKey() {
+        return fundChangeKey;
+    }
+
+    public void setFundChangeKey(WebInputLongIdKey fundChangeKey) {
+        this.fundChangeKey = fundChangeKey;
     }
 
     public BigDecimal getDelta() {
@@ -70,7 +86,8 @@ public class WebInputFundChangeUpdateInfo implements Dto {
     @Override
     public String toString() {
         return "WebInputFundChangeUpdateInfo{" +
-                "delta=" + delta +
+                "fundChangeKey=" + fundChangeKey +
+                ", delta=" + delta +
                 ", changeType='" + changeType + '\'' +
                 ", remark='" + remark + '\'' +
                 '}';