+++ familyhelper-finance-stack/src/main/java/com/dwarfeng/familyhelper/finance/stack/handler/FundChangeOperateHandler.java
@@ -19,25 +19,20 @@ public interface FundChangeOperateHandler extends Handler {
      * 记录资金变更。
      *
      * @param userKey              操作用户。
-     * @param accountBookKey       资金变更所属账本的主键。
      * @param fundChangeRecordInfo 资金变更记录信息。
      * @return 生成的资金变更实体主键。
      * @throws HandlerException 处理器异常。
      */
-    LongIdKey recordFundChange(
-            StringIdKey userKey, LongIdKey accountBookKey, FundChangeRecordInfo fundChangeRecordInfo
-    ) throws HandlerException;
+    LongIdKey recordFundChange(StringIdKey userKey, FundChangeRecordInfo fundChangeRecordInfo) throws HandlerException;
 
     /**
      * 更新资金变更。
      *
      * @param userKey              操作用户。
-     * @param fundChangeKey        资金变更实体的主键。
      * @param fundChangeUpdateInfo 资金变更更新信息。
      * @throws HandlerException 处理器异常。
      */
-    void updateFundChange(StringIdKey userKey, LongIdKey fundChangeKey, FundChangeUpdateInfo fundChangeUpdateInfo)
-            throws HandlerException;
+    void updateFundChange(StringIdKey userKey, FundChangeUpdateInfo fundChangeUpdateInfo) throws HandlerException;
 
     /**
      * 删除资金变更。